Bugfix: Campaign values not registering in GA
https://eaflood.atlassian.net/browse/IWTF-1297

In the first implementation of this plugin, we're just sending campaign
name (`cn`) and campaign medium (`cm`) values and these are not showing
in Google Analytics. Via trial and error, I've discovered that campaign
name and one of campaign source (`cs`) or campaign id (`ci`) are
required for campaign values to register. Therefore, we'll have error
feedback in hapi-gapi if the required values aren't specified.